+++ CoinSorter/FullImages.vb
@@ -1,5 +1,5 @@
 ﻿Public Class FullImageDirectory
-    Inherits Dictionary(Of Int32, FullImage)
+    Inherits Dictionary(Of int64, FullImage)
     Public Overloads Sub Add(fullImage As FullImage)
         If Not Me.ContainsKey(fullImage.ImageID) Then
             Me.Add(fullImage.ImageID, fullImage)
